Move contents to different folder
Move contents from Vectors folder to LinearAlgebra folder
