+++ Mathematics/LinearAlgebra/I2D.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Physics.NET.Mathematics.Vectors
+namespace Physics.NET.Mathematics.LinearAlgebra
 {
-    public interface IVector
+    public interface I2D
     {
     }
 }
