+++ Mathematics/LinearAlgebra/I3D.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Physics.NET.Mathematics.Vectors
+namespace Physics.NET.Mathematics.LinearAlgebra
 {
-    public interface I4D
+    public interface I3D
     {
     }
 }
