+++ Mathematics/LinearAlgebra/Polar.cs
@@ -4,12 +4,12 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Physics.NET.Mathematics.Vectors
+namespace Physics.NET.Mathematics.LinearAlgebra
 {
     /// <summary>
-    /// Polar coordinates in the form (r, θ, z).
+    /// Polar coordinates in the form (r, θ).
     /// </summary>
-    public class Cylindrical : ICoordinateSystem, I3D
+    public class Polar : ICoordinateSystem, I2D
     {
     }
 }
